pyproject.toml
[project]
name = "omniclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0,<2.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.1",
    "nanobot-ai",
    "python-dotenv>=1.1.1,<2.0.0",
    "sqlalchemy>=2.0.36,<3.0.0",
    "uvicorn>=0.30.0,<1.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
]

[tool.uv.sources]
nanobot-ai = { path = "../nanobot" }
